1.3.0-alfa1

+++ CNUnit/App.cs
@@ -1,84 +1,105 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 using System.Xml;
-using CNUnit.Tools;
 
-// ReSharper disable InconsistentNaming
-
-namespace CNUnit.Commands
+namespace CNUnit
 {
-    public class RunCommand
+    public class App
     {
-        private const int Success = 0;
-        private const int Error = 1;
-        private const int Failure = -1;
+        public const string JUnitXslt =
+            "https://raw.githubusercontent.com/nunit/nunit-transforms/master/nunit3-junit/nunit3-junit.xslt";
 
+        private static readonly string OutDirDefault = Path.Combine(Environment.CurrentDirectory, "cnunit-reports");
+        public static readonly string JUnitXsltFile = Path.Combine(Environment.CurrentDirectory, "nunit3-junit.xslt");
+        private Options Options { get; }
         private readonly List<string> _testLists = new List<string>();
 
-        public bool CNUnit_Debug;
-        public string CNUnit_Outdir = Constants.OutDirDefault;
-        public ReportType CNUnit_ReportType = ReportType.NUnit3;
-        public string CNUnit_Parse_Rules;
+        public App(Options options)
+        {
+            Options = options;
+        }
 
-        public string Tests_Dll_Path;
-        public bool Tests_Skip;
-        public bool Tests_Keep_Cases;
-        public bool Tests_Shuffle;
-
-        public string NUnit_Where;
-        public string NUnit_Executable = Utils.TryToFindNUnit();
-        public int NUnit_Workers = 1;
-        public bool NUnit_Wait;
-        public bool NUnit_No_Output;
-
-
-       
-
-
-        public int Run(string[] remainingArguments)
+        public int Run()
         {
             try
             {
-                if (!Utils.IsValidNunitConsole(NUnit_Executable))
-                {
-                    Utils.WriteLine("Error: Invalid NUnit Console executable", ConsoleColor.Red);
-                    return Error;
-                }
-
-                if (!File.Exists(Tests_Dll_Path))
-                {
-                    Utils.WriteLine("Error: Test dll is not found", ConsoleColor.Red);
-                    return Error;
-                }
-
-                Utils.OutDirSetup(CNUnit_Outdir);
-
-                if (NUnit_Wait) NUnit_No_Output = true;
-
+                TryToSetNUnitConsole();
+                CheckTestsDll();
+                SetOutDir();
                 GenerateTestLists();
-
-                if (!Tests_Skip)
-                {
-                    RunTests();
-                    if (!Tests_Keep_Cases) RemoveTestLists();
-                }
-
-                return Success;
+                RunTests();
+                RemoveTestLists();
+                return 0;
+            }
+            catch (SystemException e)
+            {
+                Utils.WriteLine(e.Message, ConsoleColor.Red);
+                return 1;
             }
             catch (Exception e)
             {
-                e.DebugCNunit(CNUnit_Debug, e.Message);
-                return Failure;
+                e.DebugCNunit(Options.IsDebug, e.Message);
+                return -1;
             }
             finally
             {
                 Utils.BeginEndPrint();
             }
+        }
+
+        private void TryToSetNUnitConsole()
+        {
+            if (Options.NUnitPath == null)
+                Options.NUnitPath = Utils.TryToFindNUnit();
+         
+            if (!Utils.IsValidNunitConsole(Options.NUnitPath))
+                throw new SystemException("Error: Invalid NUnit Console executable");
+        }
+
+        /// <summary>
+        /// Validates whether Tests DLL file is available
+        /// </summary>
+        /// <exception cref="SystemException">Error: Tests dll is not found</exception>
+        private void CheckTestsDll()
+        {
+            Console.WriteLine(Options.TestDllPath);
+            if (!File.Exists(Options.TestDllPath))
+                throw new SystemException("Error: Tests dll is not found");
+        }
+
+        /// <summary>
+        /// Validates output directory
+        /// </summary>
+        private void SetOutDir()
+        {
+            if (Options.OutDir == null)
+            {
+                Options.OutDir = OutDirDefault;
+            }
+            else
+            {
+                try
+                {
+                    Path.GetDirectoryName(Options.OutDir);
+                }
+                catch (Exception)
+                {
+                    Utils.WriteLine("Path from options is invalid. Using default one");
+                    Options.OutDir = OutDirDefault;
+                }
+            }
+
+            if (Directory.Exists(Options.OutDir))
+                Directory.Delete(Options.OutDir, true);
+            Thread.Sleep(100);
+            Directory.CreateDirectory(Options.OutDir);
+            Utils.WriteLine($"Working directory: {Options.OutDir}\n", ConsoleColor.Cyan);
         }
 
 
@@ -88,14 +109,14 @@ namespace CNUnit.Commands
         private void GenerateTestLists()
         {
             Utils.WriteLine("\nParsing NUnit tests...\n", ConsoleColor.Green);
-            var arguments = $"{NUnit_Executable} {Tests_Dll_Path} --explore:cnunit.cases;format=cases";
+            var arguments = $"{Options.NUnitPath} {Options.TestDllPath} --explore:cnunit.cases;format=cases";
 
-            if (!string.IsNullOrEmpty(NUnit_Where))
+            if (!string.IsNullOrEmpty(Options.NUnitWhere))
             {
-                arguments += $" --where \"{NUnit_Where}\"";
+                arguments += $" --where \"{Options.NUnitWhere}\"";
             }
 
-            if (CNUnit_Debug)
+            if (Options.IsDebug)
             {
                 Utils.WriteLine($"\nGenerate test list arguments: {arguments}", ConsoleColor.Blue);
             }
@@ -104,7 +125,7 @@ namespace CNUnit.Commands
             {
                 StartInfo =
                 {
-                    FileName = NUnit_Executable,
+                    FileName = Options.NUnitPath,
                     UseShellExecute = false,
                     CreateNoWindow = true,
                     Arguments = arguments
@@ -113,28 +134,29 @@ namespace CNUnit.Commands
 
             process.Start();
             process.WaitForExit();
-            var nameFile = Tests_Dll_Path.BuildTestName();
+            var nameFile = Options.TestDllPath.BuildTestName();
             var exploreResult = Path.Combine(Environment.CurrentDirectory, "cnunit.cases");
             var list = File.ReadAllLines(exploreResult).ToList();
             File.Delete(exploreResult);
 
-            if (!string.IsNullOrEmpty(CNUnit_Parse_Rules))
+            if (!string.IsNullOrEmpty(Options.CNUnitWhere))
             {
                 var parsedList = new List<string>();
-                var parseArray = CNUnit_Parse_Rules.Split(';');
+                var parseArray = Options.CNUnitWhere.Split(';');
                 foreach (var parsing in parseArray)
                 {
                     parsedList.AddRange(list.Where(v => v.Contains(parsing)));
                 }
+
                 list = parsedList;
             }
 
-            if (Tests_Shuffle) list.Shuffle();
+            if (Options.IsShuffle) list.Shuffle();
 
             Utils.WriteLine($"Total tests count: {list.Count}", ConsoleColor.Cyan);
-            Utils.WriteLine($"\nDividing tests by {NUnit_Workers} and saving to to lists:", ConsoleColor.Cyan);
+            Utils.WriteLine($"\nDividing tests by {Options.Workers} and saving to to lists:", ConsoleColor.Cyan);
             int i = 0;
-            foreach (var listPart in list.Split(NUnit_Workers))
+            foreach (var listPart in list.Split(Options.Workers))
             {
                 var fileTestBuilder = new StringBuilder();
                 var listPartElements = listPart as IList<string> ?? listPart.ToList();
@@ -142,8 +164,9 @@ namespace CNUnit.Commands
                 {
                     fileTestBuilder.AppendLine(listPartElement);
                 }
+
                 var outPut = fileTestBuilder.ToString().TrimEnd('\n', '\r');
-                var path = Path.Combine(CNUnit_Outdir, $"{nameFile}{++i}.txt");
+                var path = Path.Combine(Options.OutDir, $"{nameFile}{++i}.txt");
                 _testLists.Add(path);
                 Utils.WriteLine($"Saved {listPartElements.Count()} tests into {path}", ConsoleColor.Green);
                 File.WriteAllText(path, outPut);
@@ -156,7 +179,8 @@ namespace CNUnit.Commands
         /// </summary>
         private void RunTests()
         {
-            List<string> xmlList = new List<string>();
+            if (Options.IsGenerateOnly) return;
+            var xmlList = new List<string>();
             var allProcesses = Task.Factory.StartNew(() =>
             {
                 Utils.WriteLine("\nExecuting tests...", ConsoleColor.Cyan);
@@ -167,24 +191,27 @@ namespace CNUnit.Commands
                     {
                         StartInfo =
                         {
-                            FileName = NUnit_Executable,
+                            FileName = Options.NUnitPath,
                             Arguments = BuildNUnitTestListArguments(test)
                         }
                     };
 
-                    pr.StartInfo.UseShellExecute = NUnit_No_Output;
+                    pr.StartInfo.UseShellExecute = Options.IsNUnitWait;
                     pr.Start();
-                    if (CNUnit_Debug) Utils.WriteLine("CNUnit: Run tests params " + pr.StartInfo.Arguments, ConsoleColor.Blue);
+                    if (Options.IsDebug)
+                        Utils.WriteLine("CNUnit: Run tests params " + pr.StartInfo.Arguments, ConsoleColor.Blue);
                     processes.Add(pr);
                 }
+
                 foreach (var pr in processes)
                 {
                     pr.WaitForExit();
                     pr.Dispose();
                 }
+
                 foreach (var test in _testLists)
                 {
-                    var path = Path.Combine(CNUnit_Outdir, test.BuildTestName() + ".xml");
+                    var path = Path.Combine(Options.OutDir, test.BuildTestName() + ".xml");
                     xmlList.Add(path);
                     Utils.WriteLine($"Test report saved - {path}", ConsoleColor.Green);
                 }
@@ -201,26 +228,27 @@ namespace CNUnit.Commands
         private string BuildNUnitTestListArguments(string test)
         {
             var sb = new StringBuilder();
-            sb.Append(Tests_Dll_Path);
+            sb.Append(Options.TestDllPath);
             sb.Append($" --testlist=\"{test}\"");
 
-            switch (CNUnit_ReportType)
+            switch (Options.OutputFortmat)
             {
                 case ReportType.Nunit2:
                     sb.Append($" --result=\"{test.BuildTestName()}.xml;format=nunit2\" ");
                     break;
                 case ReportType.JUnit:
-                    if (!File.Exists(Constants.JUnitXsltFile))
+                    if (!File.Exists(JUnitXsltFile))
                         Utils.DownloadJUnitTransform();
                     sb.Append(
-                        $" --result=\"{test.BuildTestName()}.xml\";transform=\"{Constants.JUnitXsltFile}\" ");
+                        $" --result=\"{test.BuildTestName()}.xml\";transform=\"{JUnitXsltFile}\" ");
                     break;
                 default:
                     sb.Append($" --result=\"{test.BuildTestName()}.xml\" ");
                     break;
             }
-            if (!string.IsNullOrEmpty(CNUnit_Outdir)) sb.Append($" --work=\"{CNUnit_Outdir}\"").Append(" ");
-            if (NUnit_Wait) sb.Append(" --wait ");        
+
+            if (!string.IsNullOrEmpty(Options.OutDir)) sb.Append($" --work=\"{Options.OutDir}\"").Append(" ");
+            if (Options.IsNUnitWait) sb.Append(" --wait ");
             return sb.ToString();
         }
 
@@ -230,26 +258,27 @@ namespace CNUnit.Commands
         /// </summary>
         private void RemoveTestLists()
         {
-            if (CNUnit_Debug)
+            if (Options.IsGenerateOnly) return;
+            if (Options.IsKeepTestCases) return;
+            if (Options.IsDebug)
                 Utils.WriteLine("\nRemoving test lists:", ConsoleColor.Blue);
             foreach (var testList in _testLists)
             {
                 File.Delete(testList);
-                if (CNUnit_Debug)
+                if (Options.IsDebug)
                     Utils.WriteLine($"Removed {testList}", ConsoleColor.DarkBlue);
             }
         }
 
-        private void NUnitXmlUpdate(List<string> xmls)
+        private void NUnitXmlUpdate(IEnumerable<string> xmls)
         {
             try
             {
                 foreach (var xml in xmls)
                 {
-                    Console.WriteLine("!!!!");
                     var doc = new XmlDocument();
                     doc.LoadXml(File.ReadAllText(xml));
-                    if (CNUnit_ReportType == ReportType.JUnit)
+                    if (Options.OutputFortmat == ReportType.JUnit)
                     {
                         var testSuiteNodeList = doc.DocumentElement?.SelectNodes("//testsuite");
                         if (testSuiteNodeList != null)
@@ -257,8 +286,8 @@ namespace CNUnit.Commands
                             {
                                 if (testSuiteNode.Attributes != null)
                                     testSuiteNode.Attributes["name"].Value =
-                                        Path.GetFileNameWithoutExtension(xml)+ "."+ Path.GetRandomFileName();
-                            } 
+                                        Path.GetFileNameWithoutExtension(xml) + "." + Path.GetRandomFileName();
+                            }
                     }
                     else
                     {
@@ -269,14 +298,14 @@ namespace CNUnit.Commands
                             xmlAttributeCollection.Value = Path.GetFileNameWithoutExtension(xml);
                         }
                     }
-                
+
                     doc.Save(xml);
                 }
             }
             catch (Exception e)
             {
-                e.DebugCNunit(CNUnit_Debug, "Unable to parse NUnit xml results");
-            }    
+                e.DebugCNunit(Options.IsDebug, "Unable to parse NUnit xml results");
+            }
         }
     }
 }
